requirement.txt
pytest
sendgrid
behave-html-formatter
selenium
logzero